Refactor to move ActionProvider functionality into ActionsViewTable and remove unused imports and remove author
